pyproject.toml
[project]
name = "anacreon"
version = "0.1.0"
description = "Anacreon is a django project to help small business, track inventory, capital, sales, profits, losses etc"
python = ">=3.12"
authors = [
    {name = "Keeya Emmanuel",email = "ekeeya@thothcode.tech"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
anyio = "4.10.0"
asgiref = "3.9.1"
build = "1.3.0"
cachecontrol = "0.14.3"
certifi = "2025.8.3"
cffi = "1.17.1"
charset-normalizer = "3.4.3"
cleo = "2.1.0"
crashtest = "0.4.1"
distlib = "0.4.0"
django = "5.2.5"
dulwich = "0.22.8"
fastjsonschema = "2.21.2"
filelock = "3.19.1"
findpython = "0.6.3"
h11 = "0.16.0"
httpcore = "1.0.9"
httpx = "0.28.1"
idna = "3.10"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jaraco-context = "6.0.1"
jaraco-functools = "4.2.1"
keyring = "25.6.0"
more-itertools = "10.7.0"
msgpack = "1.1.1"
packaging = "25.0"
pbs-installer = "2025.8.14"
pkginfo = "1.12.1.2"
platformdirs = "4.3.8"
poetry = "2.1.4"
poetry-core = "2.1.3"
pycparser = "2.22"
pyproject-hooks = "1.2.0"
rapidfuzz = "3.13.0"
requests = "2.32.4"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
sniffio = "1.3.1"
sqlparse = "0.5.3"
tomlkit = "0.13.3"
trove-classifiers = "2025.8.6.13"
typing-extensions = "4.14.1"
urllib3 = "2.5.0"
virtualenv = "20.32.0"
xattr = "1.2.0"
zstandard = "0.23.0"
psycopg2 = "^2.9.10"
djangorestframework = "^3.16.1"
django-filter = "^25.1"
markdown = "^3.8.2"
iso8601 = "^2.1.0"
pandas = "^2.3.1"
xlrd = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false
